requirements.txt
streamlit==1.27.0
anthropic==0.5.0
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
feedparser==6.0.10
pillow>=7.1.0,<10
arxiv==1.4.8
apscheduler==3.10.1
google-api-python-client==2.100.0
urllib3==2.0.4
pytest==7.4.2
pytest-mock==3.11.1